Poprawki dotyczące wyglądu kodu oraz poprawka dotyczaca logowania.

+++ src/main/java/pl/reportsController/users/UserController.java
@@ -1,13 +1,10 @@
 package pl.reportsController.users;
 
-import jakarta.servlet.ServletOutputStream;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
-import org.apache.catalina.manager.HTMLManagerServlet;
 import org.springframework.web.bind.annotation.*;
-import pl.reportsController.passwords.PasswordHashing;
 
-import java.util.List;
+import java.util.HashMap;
 
 @RestController
 @RequestMapping("/users")
@@ -20,61 +17,78 @@ public class UserController {
     }
 
     @GetMapping
-    public Iterable<UserEntity> getAllUsers(){
+    public Iterable<UserEntity> getAllUsers() {
         return userRepository.findAll();
     }
 
     @GetMapping("/{id}")
-    public UserEntity getUserById(@PathVariable Long id){
+    public UserEntity getUserById(
+            @PathVariable Long id) {
         return userRepository.findById(id).orElseThrow(RuntimeException::new);
     }
 
     @PostMapping("/login")
-    public String loginUser(HttpServletResponse response, @RequestBody UserEntity userEntity) {
+    public String loginUser(HttpServletResponse response,
+                            @RequestBody UserEntity userEntity) {
         String hashedPassword = userEntity.getPassword();
 
-        if(userRepository.findByLogin(userEntity.getLogin()) == null ){
-            System.out.println("Error login (wrong username): "+userEntity.getLogin());
+        if (userRepository.findByLogin(userEntity.getLogin()) == null) {
+            System.out.println("Error login (wrong username): " + userEntity.getLogin());
             return "Błędny login";
         }
 
-        if(!userRepository.findPasswordByLogin(userEntity.getLogin()).equals(hashedPassword)){
-            System.out.println("Error login (wrong password): "+userEntity.getPassword());
+        if (!userRepository.findPasswordByLogin(userEntity.getLogin()).equals(hashedPassword)) {
+            System.out.println("Error login (wrong password): " + userEntity.getLogin());
             return "Błędne hasło";
         }
         Long userId = userRepository.findIdByUsernameAndPassword(userEntity.getLogin(), hashedPassword);
         System.out.println(userEntity.getLogin() + " has been logged in to service!");
-        Cookie cookie = new Cookie("userId", ""+userId);
-        cookie.setHttpOnly(true);
 
-        response.addCookie(cookie);
-
+        HashMap<String, String> cookies = new HashMap<>();
+        cookies.put("isLogged", "true");
+        cookies.put("userId", userId.toString());
+        Cookie cookie = null;
+        for (String name : cookies.keySet()) {
+            cookie = new Cookie(name, cookies.get(name));
+            response.addCookie(cookie);
+        }
         return "OK";
     }
 
     @PostMapping("/register")
-    public String registerUser(HttpServletResponse response, @RequestBody UserEntity ue ){
+    public String registerUser(HttpServletResponse response,
+                               @RequestBody UserEntity ue) {
 
         UserEntity user = new UserEntity(ue.getLogin(), ue.getPassword(), ue.getEmail());
 
-        if(userRepository.checkLoginExists(ue.getLogin()) != null){
+        System.out.println(user.toString());
+        if (userRepository.checkLoginExists(ue.getLogin()) != null) {
             return "Login jest już zajęty";
         }
 
-        if(userRepository.checkEmailExists(ue.getEmail()) != null){
+        if (userRepository.checkEmailExists(ue.getEmail()) != null) {
             return "Email jest już zajęty";
         }
 
-         userRepository.save(ue);
+        userRepository.save(user);
 
-        Cookie cookie = new Cookie("idUser", ""+ue.getId());
-        response.addCookie(cookie);
+        HashMap<String, String> cookies = new HashMap<>();
+        cookies.put("isLogged", "true");
+        cookies.put("userId", user.getId().toString());
+        for (String name : cookies.keySet()) {
+            response.addCookie(new Cookie(name, cookies.get(name)));
+        }
         return "OK";
     }
 
     @DeleteMapping("delete={id}")
-    public void deleteUserById(@PathVariable Long id){
-        if(userRepository.existsById(id)){
+    public void deleteUserById(
+            @RequestBody UserEntity ue,
+            @PathVariable Long id) {
+        if (ue.getUserRole() != UserRole.ADMINISTRATOR) {
+            return;
+        }
+        if (userRepository.existsById(id)) {
             throw new RuntimeException();
         } else {
             userRepository.deleteById(id);
